Server Ports Analysis Page - table view refactor (#8133) (#8150)
* Fix discrepancy between table view and chart view in server ports analysis page (#8133).

* Fix to handle only the master L7 protocol.

+++ include/HostsPortsAnalysis.h
@@ -1,6 +1,6 @@
 /*
  *
- * (C) 2013-23 - ntop.org
+ * (C) 2013-24 - ntop.org
  *
  *
  * This program is free software; you can redistribute it and/or modify
@@ -29,7 +29,7 @@ class HostsPortsAnalysis {
     private:
         u_int16_t port = 0;
         int l7_protocol_id = 0;
-        u_int32_t l4_protocol_id = 0;
+        u_int8_t l4_protocol_id = 0;
         std::unordered_map<u_int64_t, HostDetails *> *hosts_details;
     
     public:
@@ -49,13 +49,13 @@ class HostsPortsAnalysis {
         inline u_int16_t get_port() { return(port); };
         inline std::unordered_map<u_int64_t, HostDetails *>* get_hosts_details() { return(hosts_details); };
         inline int get_l7_proto() { return(l7_protocol_id); };
-        inline u_int32_t get_l4_proto() { return(l4_protocol_id); };
+        inline u_int8_t get_l4_proto() { return(l4_protocol_id); };
         
         /* Setters */
         void add_host_details(HostDetails *host_details);
         inline void set_port(u_int16_t _port) { port = _port; };
         inline void set_l7_proto(int l7_proto) { l7_protocol_id = l7_proto; };
-        inline void set_l4_proto(u_int32_t l4_proto) { l4_protocol_id = l4_proto; };
+        inline void set_l4_proto(u_int8_t l4_proto) { l4_protocol_id = l4_proto; };
 };
 
 #endif /* _HOSTS_PORTS_ANALYSIS_H_ */
